Improved no_duplicate plugin to update tags when deduplicating.
The mail_id column is added and set in a postprocess plugin.
When a duplicate is discarded, the tags it could have
at the point of the discard can be optionally assigned
to the previous instance of the same mail.
This is primarily useful to track messages across folder moves,
when manitou-mail is used to permanently sync from an IMAP server.

Also, add a trigger to delete from no_duplicate_import when mail
is deleted.